// src/app/interfaces/personaMoral.ts
export interface IPersonaMoral {
    idPersonaMoral: number;
    razonSocial: string;
    razonComercial: string;
    fechaConstitucion: string;
    rfc: string;
    nacionalidad: string;
    paisRegistro: string;
    estadoRegistro: string;
    ciudadRegistro: string;
    numEscritura: string;
    fechaRppc: string;
    nombreNotario: string;
    numNotario: string;
    folioMercantil: string;
    calle: string;
    numExterior: string;
    numInterior?: string;
    colonia: string;
    codigoPostal: string;
    paisResidencia: string;
    estadoResidencia: string;
    ciudadResidencia: string;
}